Refctor: organize config_parser by responsibility (parser/validation)
